Renamed files in offst-bin (filename -> filenamelib).
