+++ components/bin/src/bin/stindex.rs
@@ -16,7 +16,7 @@ extern crate log;
 
 use structopt::StructOpt;
 
-use bin::stindex::{stindex, IndexServerBinError, StIndexCmd};
+use bin::stindexlib::{stindex, IndexServerBinError, StIndexCmd};
 
 fn run() -> Result<(), IndexServerBinError> {
     env_logger::init();
