+++ components/bin/src/stindexlib.rs
