+++ components/bin/src/stmgrlib.rs
