+++ components/bin/src/stnodelib.rs
