+++ components/bin/src/strelaylib.rs
